third_party/WebKit/LayoutTests/svg/dom/SVGLengthList-setEmptyString-expected.txt

PASS text.x.baseVal.numberOfItems is 3
PASS text.x.baseVal.numberOfItems is 0
PASS text.x.baseVal.numberOfItems is 3
PASS text.x.baseVal.numberOfItems is 0
PASS successfullyParsed is true

TEST COMPLETE
